Crate order_and_chaos

Structs

Game
A game of order and chaos.

Move
A move consists of a piece and a location in the board.

Enums

GameStatus
A game of Order and Chaos may be in progress, or won by either player.

MoveType
A move in Order and Chaos is either placing an X piece or an O piece.

Player
A Player is either Order or Chaos.

Traits

Strategy
Provides the strategy for an AI player to play the game. A new AI player can be implemented simply by defining a new implementation for ai_move.