Module order_and_chaos::board

A Board defines a game of Order and Chaos. A user facing client application should utilize the enums and structs declared publicly in this module. A game is considered to be an immutable structure. Each player's move will result in a new game with the correct state.

To use this as a library, a client application will need to create an initial game object and manage any relevant input or output data. The GameStatus enum will inform the client whether or not the game has been won by any player. A formatter for println! is provided within the game struct; however, any other UI for the game is the responsibility of the client.

Structs

Game
A game of order and chaos.

Move
A move consists of a piece and a location in the board.

Enums

BoardDirection
Defines directions for checking if the game is in a win state.

GameStatus
A game of Order and Chaos may be in progress, or won by either player.

MoveType
A move in Order and Chaos is either placing an X piece or an O piece.

Player
A Player is either Order or Chaos

Traits

Strategy
Provides the strategy for an AI player to play the game. A new AI player can be implemented simply by defining a new implementation for ai_move.